+++ notifier/src/main/java/io/reflectoring/jiraalerts/notify/NotifyDeviceJob.java
@@ -1,4 +1,4 @@
-package io.reflectoring.jiraalerts.issuerecognition;
+package io.reflectoring.jiraalerts.notify;
 
 import org.quartz.JobExecutionContext;
 import org.quartz.JobExecutionException;
@@ -10,13 +10,13 @@ import org.springframework.stereotype.Component;
  * Job that encapsulates the recognition of new issues and the trigger on the reaction.
  */
 @Component
-public class IssueRecognitionJob extends QuartzJobBean {
+public class NotifyDeviceJob extends QuartzJobBean {
 
 	@Autowired
-	private IssueRecognizerService issueRecognizerService;
+	private NotifyDeviceService issueRecognizerService;
 
 	@Override
 	protected void executeInternal(JobExecutionContext context) throws JobExecutionException {
-		issueRecognizerService.recognize();
+        issueRecognizerService.notifyDevices();
 	}
 }
