+++ settings.gradle
@@ -1,2 +1,3 @@
 include 'jira-rest-client'
 include 'persistence'
+include 'notifier'
