+++ src/main/java/io/reflectoring/jiraalerts/application/spring/JiraAlertsConfiguration.java
@@ -10,6 +10,7 @@ import org.springframework.boot.web.servlet.ServletContextInitializer;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 
+import io.reflectoring.jiraalerts.quartz.QuartzTimerConfiguration;
 import io.reflectoring.jiraalerts.spring.PersistenceContextConfiguration;
 
 /**
